Support for multiple sparklines with different data in one graph
To render multiple curves with different data, provide an array of
arrays of values as data. data[0] will be used as data for the first
child, data[1] for the next child and so forth. If data is not an array
of arrays, the behavior remains unchanged (i.e. all curves will use the
same data).